+++ Classes/Instapaper/InstapaperRequest.m
@@ -54,7 +54,7 @@
 
 - (void)connection:(NSURLConnection *)connection didReceiveResponse:(NSURLResponse *)response {
     if ([response isKindOfClass:[NSHTTPURLResponse class]]) {
-        int status = [(NSHTTPURLResponse *) response statusCode];
+        NSInteger status = [(NSHTTPURLResponse *) response statusCode];
         if (status == 403) [self failWithErrorText:@"Invalid username or password."];
         if (status == 500) [self failWithErrorText:@"Internal error, try again later."];
         if (status == 201) [self succeed];
