+++ Classes/XML/XMLDocument.m
@@ -59,7 +59,7 @@
     if (nodes == NULL) return nil;
     
     NSMutableArray *result = [NSMutableArray array];
-    for (int i = 0; i < nodes->nodeNr; i++) {
+    for (NSInteger i = 0; i < nodes->nodeNr; i++) {
         XMLElement *element = [[XMLElement alloc] initWithNode:nodes->nodeTab[i] inDocument:self];
         [result addObject:[element autorelease]];
     }
